# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
...
paths:
...
  /api/acp/registry:
    get:
      operationId: listAcpRegistry
      summary: List agents in the ACP registry
      responses:
        "200":
          description: Registry entries
    post:
      operationId: registerAcpAgent
      summary: Register an agent in the ACP registry
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Registered
  /api/acp/install:
    post:
      operationId: installAcpAgent
      summary: Install an ACP agent
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Installed
    delete:
      operationId: uninstallAcpAgent
      summary: Uninstall an ACP agent
      responses:
        "200":
          description: Uninstalled
...